tests/spatnav-scroll-001.html

<!doctype html>
<html lang=en>
<meta charset="utf-8">
<title>Spatnav sanity check</title>
  <link rel="author" title="Florian Rivoal" href="https://florian.rivoal.net">
  <meta name=flags content="">
  <link rel=help href="https://wicg.github.io/spatial-navigation/#high-level-api">

  <script src="../resources/testharness.js"></script>
  <script src="../resources/testharnessreport.js"></script>

<!--load polyfill. To be removed before submitting to WPT-->
<script src="../polyfill/spatnav-heuristic.js"></script>
<script src="../polyfill/spatnav-api.js"></script>
<script>focusNavigationHeuristics({"standardName": "true" });</script>

<style>
#d {
  width: 100px;
  height: 100px;
  overflow: scroll;
}
</style>

<body>
  <div id=d tabindex=0>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras lacinia nulla justo, et lobortis odio condimentum facilisis. Cras tempor pulvinar ligula finibus vulputate. Vivamus posuere accumsan metus, eget consectetur diam. Donec convallis, ipsum sed semper tempus, elit quam luctus tortor, in molestie felis ex eu leo. Suspendisse pellentesque non diam sit amet dictum. Cras nibh ex, ultricies non lorem vitae, placerat euismod sem. Quisque nec luctus risus, vel tincidunt eros. Mauris tempor enim et mi sollicitudin blandit.</div>
</body>

<script async>
test(function() {
  let d = document.getElementById("d");
  d.focus();
  d.scrollTop = 0;
  window.navigate("down");
  assert_not_equals(d.scrollTop, 0);
}, "spatnav on a scrollable element with no focusable descendant causes scrolling")
</script>
